@@ -1,18 +1,17 @@
 name: flutter_midi_command
 description: A Flutter plugin for sending and receiving MIDI messages between Flutter and physical and virtual MIDI devices. Wraps CoreMIDI and android.media.midi in a thin dart/flutter layer.
-version: 0.2.6
+version: 0.3.0
 homepage: https://github.com/InvisibleWrench/FlutterMidiCommand
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_midi_command_platform_interface: ^0.2.0
-#  flutter_midi_command_linux: ^0.0.1
-
+  flutter_midi_command_platform_interface: ^0.3.1
+  flutter_midi_command_linux: ^0.1.0
 
 dev_dependencies:
   flutter_test:
@@ -36,8 +35,8 @@ flutter:
         pluginClass: FlutterMidiCommandPlugin
       macos:
         pluginClass: SwiftFlutterMidiCommandPlugin
-#      linux:
-#        default_package: flutter_midi_command_linux
+      linux:
+        default_package: flutter_midi_command_linux
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
